fix dtypes

+++ tests/nightly/test_large_vector.py
@@ -346,7 +346,7 @@ def test_sequence_reverse():
 
 
 def test_sequence_last():
-    a = nd.arange(0, LARGE_X * 2).reshape(LARGE_X, 2)
+    a = nd.arange(0, LARGE_X * 2, dtype="int64").reshape(LARGE_X, 2)
 
     # test if returns last sequence
     b = nd.SequenceLast(a)
@@ -356,7 +356,7 @@ def test_sequence_last():
     # test with sequence length
     # parameter sequence_length - NDArray with shape (batch_size)
     # (2,3) indicates 2nd sequence from batch 1 and 3rd sequence from batch 2
-    b = nd.SequenceLast(a, sequence_length=mx.nd.array([2, 3]),
+    b = nd.SequenceLast(a, sequence_length=mx.nd.array([2, 3], dtype="int64"),
                         use_sequence_length=True)
     # check if it takes 2nd sequence from the first batch
     assert b[0] == a[1][0]
